+++ src/main/java/utp/restaurant/dao/RoleDAO.java
@@ -64,7 +64,7 @@ public class RoleDAO implements DAO<Role> {
         try {
 
             ps = conn.prepareStatement(query);
-            ps.setLong(1, id);
+            ps.setInt(1, (int) id);
 
             rs = ps.executeQuery();
